docs: translate all JavaDocs in Server module * refactor: cleanup more code

+++ Allay-Server/src/main/java/org/allaymc/server/client/storage/AllayNBTFilePlayerStorage.java
@@ -8,7 +8,6 @@ import org.allaymc.api.client.storage.PlayerData;
 import org.cloudburstmc.nbt.NbtMap;
 import org.cloudburstmc.nbt.NbtUtils;
 
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.UUID;
@@ -34,10 +33,11 @@ public class AllayNBTFilePlayerStorage implements NativeFilePlayerStorage {
     public PlayerData readPlayerData(UUID uuid) {
         var path = buildPlayerDataFilePath(uuid);
         if (!Files.exists(path)) return PlayerData.createEmpty();
+
         try (var reader = NbtUtils.createGZIPReader(Files.newInputStream(path))) {
             return PlayerData.fromNBT((NbtMap) reader.readTag());
         } catch (Throwable e) {
-            log.error("Error while reading player data " + uuid, e);
+            log.error("Error while reading player data {}", uuid, e);
             return PlayerData.createEmpty();
         }
     }
@@ -46,23 +46,25 @@ public class AllayNBTFilePlayerStorage implements NativeFilePlayerStorage {
     @Override
     public void savePlayerData(UUID uuid, PlayerData playerData) {
         var path = buildPlayerDataFilePath(uuid);
+
         var oldPath = path.resolveSibling(uuid + "_old.nbt");
         if (Files.exists(oldPath)) {
             // The old file
-            log.warn("Undeleted tmp player data file is found, which may caused by incorrect shutdown. File: " + oldPath);
+            log.warn("Undeleted tmp player data file is found, which may caused by incorrect shutdown. File: {}", oldPath);
             Files.delete(oldPath);
         }
-        if (Files.exists(path)) {
-            // rename current file to uuid_old.nbt
-            Files.move(path, oldPath);
-        }
+
+        // rename current file to uuid_old.nbt
+        if (Files.exists(path)) Files.move(path, oldPath);
+
         try (var writer = NbtUtils.createGZIPWriter(Files.newOutputStream(path))) {
             writer.writeTag(playerData.toNBT());
         } catch (Throwable e) {
             // error, rename uuid_old.nbt file to uuid.nbt
             Files.move(oldPath, path);
-            log.error("Error while writing player data " + uuid, e);
+            log.error("Error while writing player data {}", uuid, e);
         }
+
         // delete uuid_old.nbt file
         Files.deleteIfExists(oldPath);
     }
