+++ Allay-Server/src/main/java/org/allaymc/server/utils/ComponentClassCacheUtils.java
@@ -9,7 +9,6 @@ import org.allaymc.api.utils.JSONUtils;
 import org.allaymc.server.Allay;
 import org.apache.commons.io.file.PathUtils;
 
-import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -27,15 +26,15 @@ import java.util.Properties;
 @SuppressWarnings("unchecked")
 public final class ComponentClassCacheUtils {
     public static final Path CACHE_ROOT_PATH = Path.of("caches");
+
     private static final Object2ObjectOpenHashMap<String, String> CACHE_MAP = new Object2ObjectOpenHashMap<>();
 
     static {
-        File cacheRootDir = CACHE_ROOT_PATH.toFile();
-        if (!cacheRootDir.exists()) {
-            if (!cacheRootDir.mkdir()) {
-                throw new RuntimeException("Can't create cache directory!");
-            }
+        var cacheRootDir = CACHE_ROOT_PATH.toFile();
+        if (!cacheRootDir.exists() && !cacheRootDir.mkdir()) {
+            throw new RuntimeException("Can't create cache directory!");
         }
+
         try {
             Allay.EXTRA_RESOURCE_CLASS_LOADER.addURL(CACHE_ROOT_PATH.toUri().toURL());
         } catch (MalformedURLException e) {
@@ -47,21 +46,21 @@ public final class ComponentClassCacheUtils {
      * Check cache valid,Every time the git commit number changes, the cache will become invalid.
      */
     public static void checkCacheValid() {
-        Path cacheValid = CACHE_ROOT_PATH.resolve("cache.valid");
-        Properties properties = new Properties();
+        var cacheValid = CACHE_ROOT_PATH.resolve("cache.valid");
+        var properties = new Properties();
         try (var input = new InputStreamReader(Objects.requireNonNull(Allay.EXTRA_RESOURCE_CLASS_LOADER.getResourceAsStream("git.properties")))) {
             properties.load(input);
             if (Files.exists(cacheValid) &&
-                    Files.readString(cacheValid).equals(properties.getProperty("git.commit.id.abbrev")) &&
-                    CACHE_ROOT_PATH.resolve("mapping.json").toFile().exists()
+                Files.readString(cacheValid).equals(properties.getProperty("git.commit.id.abbrev")) &&
+                CACHE_ROOT_PATH.resolve("mapping.json").toFile().exists()
             ) {
                 log.info(I18n.get().tr(TrKeys.A_CACHE_LOAD));
                 return;
             }
+
             var cn = CACHE_ROOT_PATH.resolve("cn");
-            if (Files.exists(cn)) {
-                PathUtils.deleteDirectory(cn);
-            }
+            if (Files.exists(cn)) PathUtils.deleteDirectory(cn);
+
             Files.deleteIfExists(CACHE_ROOT_PATH.resolve("mapping.json"));
         } catch (IOException e) {
             throw new RuntimeException(e);
@@ -69,8 +68,7 @@ public final class ComponentClassCacheUtils {
             try {
                 Files.deleteIfExists(CACHE_ROOT_PATH.resolve("cache.valid"));
                 Files.writeString(CACHE_ROOT_PATH.resolve("cache.valid"), properties.getProperty("git.commit.id.abbrev"), StandardCharsets.UTF_8, StandardOpenOption.CREATE);
-            } catch (IOException ignore) {
-            }
+            } catch (IOException ignore) {}
         }
     }
 
@@ -80,7 +78,7 @@ public final class ComponentClassCacheUtils {
 
     public static void saveCacheMapping() {
         try {
-            Path resolve = CACHE_ROOT_PATH.resolve("mapping.json");
+            var resolve = CACHE_ROOT_PATH.resolve("mapping.json");
             Files.deleteIfExists(resolve);
             Files.writeString(resolve, JSONUtils.to(CACHE_MAP), StandardCharsets.UTF_8, StandardOpenOption.CREATE);
         } catch (IOException e) {
@@ -89,26 +87,24 @@ public final class ComponentClassCacheUtils {
     }
 
     public static void readCacheMapping() {
-        File file = CACHE_ROOT_PATH.resolve("mapping.json").toFile();
-        if (file.exists()) {
-            try (var reader = new FileReader(file)) {
-                CACHE_MAP.putAll(JSONUtils.from(reader, Map.class));
-                CACHE_MAP.trim();
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
+        var file = CACHE_ROOT_PATH.resolve("mapping.json").toFile();
+        if (!file.exists()) return;
+
+        try (var reader = new FileReader(file)) {
+            CACHE_MAP.putAll(JSONUtils.from(reader, Map.class));
+            CACHE_MAP.trim();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
         }
     }
 
     public static <T> Class<T> getCacheClass(Class<T> interfaceClass) {
-        String s = CACHE_MAP.get(interfaceClass.getSimpleName());
+        var s = CACHE_MAP.get(interfaceClass.getSimpleName());
         try {
-            if (s == null) {
-                return null;
-            }
+            if (s == null) return null;
             return (Class<T>) Allay.EXTRA_RESOURCE_CLASS_LOADER.loadClass(s);
         } catch (ClassNotFoundException ignore) {
-            CACHE_MAP.remove(s);//remove old cache entry
+            CACHE_MAP.remove(s); // remove old cache entry
             return null;
         }
     }
